Parser: report source, fix a bug on empty lines
When we log a message in the parser we now log what the source of the
file is. This is very useful when use several block list sources and
there is an error in one of them.

An empty line no longer breaks the parser, and they are now silently
ignored.