Allow any ip address in block lists
I chose to only allow 0.0.0.0 / :: as the ip as the default block list
for pi-hole uses that except for some standard hostnames such as
localhost etc:
https://raw.githubusercontent.com/StevenBlack/hosts/master/hosts

However, I found other blocklists use 127.0.0.1 as the ip address so
they would not be effective at all with that assumption. Blocking
queries for "localhost" should likely be okay - you would not hit the
resolver if it's in your /etc/hosts anyway.